Extract SyncRequestMaker from Worker
